code/test/requirements.txt
Flask==0.12.1
google-api-python-client==1.6.2
face_recognition 
pickle
codecs
numpy
PIL
json